pyproject.toml
[project]
name = "PDF-Kacper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "numpy",
    "matplotlib",
    "scikit-image",
    "reportlab",
    "pypdf",
    "pymupdf",
    "pillow",
    "tkinter",
]

[dependency-groups]
dev = ["ipykernel", "ruff", "pyinstaller", "pillow"]
